BUG FIX: In integer/long to BigDecimal, don't convert the number to a double first.
git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@917364 13f79535-47bb-0310-9956-ffa450edef68

+++ framework/base/src/org/ofbiz/base/conversion/NumberConverters.java
@@ -473,7 +473,7 @@ public class NumberConverters implements ConverterLoader {
         }
 
         public BigDecimal convert(Integer obj) throws ConversionException {
-            return BigDecimal.valueOf(obj.doubleValue());
+            return BigDecimal.valueOf(obj.intValue());
         }
     }
 
@@ -572,7 +572,7 @@ public class NumberConverters implements ConverterLoader {
         }
 
         public BigDecimal convert(Long obj) throws ConversionException {
-            return BigDecimal.valueOf(obj.doubleValue());
+            return BigDecimal.valueOf(obj.longValue());
         }
     }
